Drop glm support, use opengl functions directly instead

+++ yaf_opengl_engine/src/YafMisc.cpp
@@ -20,8 +20,8 @@ YafXYZ::YafXYZ(const std::string& s)
 
 YafXYZ YafXYZ::GetNormalized()
 {
-        double length = sqrt(X * X + Y * Y + Z * Z);
-        return YafXYZ(X / length, Y / length, Z / length);
+    float length = sqrt(X * X + Y * Y + Z * Z);
+    return YafXYZ(X / length, Y / length, Z / length);
 }
